+++ NEWS.md
@@ -9,6 +9,8 @@
 * Add `ga_model_tweet` to enable data viz in tweets
 * Add `ga_model_write` to write model functions to a file
 * Add userActivity API via `gar_clientid_activity()`
+* Use new authentication setup via `googleAuthR v0.9.0` and `gargle`
+* Prevent use of default client id project in non-interactive scripts.  
 
 # googleAnalyticsR 0.6.0